+++ crates/rust-client/src/rpc/endpoint.rs
@@ -1,9 +1,5 @@
-//! Provides code related to configuring the client.
-
 use alloc::string::{String, ToString};
-use core::fmt::{self, Debug};
-
-use serde::{Deserialize, Serialize};
+use core::fmt;
 
 // ENDPOINT
 // ================================================================================================
@@ -12,7 +8,7 @@ use serde::{Deserialize, Serialize};
 /// port.
 ///
 /// This struct is used to define the address of a Miden node that the client will connect to.
-#[derive(Clone, Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
+#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
 pub struct Endpoint {
     /// The protocol used to connect to the endpoint (e.g., "http", "https").
     protocol: String,
@@ -23,6 +19,8 @@ pub struct Endpoint {
 }
 
 impl Endpoint {
+    pub(crate) const MIDEN_NODE_PORT: u16 = 57291;
+
     /// Creates a new `Endpoint` with the specified protocol, host, and port.
     ///
     /// # Arguments
@@ -33,9 +31,7 @@ impl Endpoint {
     pub const fn new(protocol: String, host: String, port: u16) -> Self {
         Self { protocol, host, port }
     }
-}
 
-impl Endpoint {
     pub fn protocol(&self) -> &str {
         &self.protocol
     }
@@ -55,13 +51,12 @@ impl fmt::Display for Endpoint {
     }
 }
 
-const MIDEN_NODE_PORT: u16 = 57291;
 impl Default for Endpoint {
     fn default() -> Self {
         Self {
             protocol: "http".to_string(),
             host: "localhost".to_string(),
-            port: MIDEN_NODE_PORT,
+            port: Self::MIDEN_NODE_PORT,
         }
     }
 }
@@ -97,7 +92,7 @@ impl TryFrom<&str> for Endpoint {
                 // skip the separator
                 let hostname = &hostname[3..];
 
-                (protocol, hostname, MIDEN_NODE_PORT)
+                (protocol, hostname, Self::MIDEN_NODE_PORT)
             },
             (None, Some(port_idx)) => {
                 let (hostname, port) = endpoint.split_at(port_idx);
@@ -105,7 +100,7 @@ impl TryFrom<&str> for Endpoint {
 
                 ("https", hostname, port)
             },
-            (None, None) => ("https", endpoint, MIDEN_NODE_PORT),
+            (None, None) => ("https", endpoint, Self::MIDEN_NODE_PORT),
         };
 
         Ok(Endpoint {
@@ -116,37 +111,11 @@ impl TryFrom<&str> for Endpoint {
     }
 }
 
-// RPC CONFIG
-// ================================================================================================
-
-/// Settings for the RPC client.
-#[derive(Debug, Deserialize, Eq, PartialEq, Serialize)]
-pub struct RpcConfig {
-    /// Address of the Miden node to connect to.
-    pub endpoint: Endpoint,
-    /// Timeout for the RPC api requests, in milliseconds.
-    #[serde(default = "default_timeout")]
-    pub timeout_ms: u64,
-}
-
-const fn default_timeout() -> u64 {
-    10000
-}
-
-impl Default for RpcConfig {
-    fn default() -> Self {
-        Self {
-            endpoint: Endpoint::default(),
-            timeout_ms: 10000,
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
     use alloc::string::ToString;
 
-    use crate::config::{Endpoint, MIDEN_NODE_PORT};
+    use crate::rpc::Endpoint;
 
     #[test]
     fn test_endpoint_parsing_with_hostname_only() {
@@ -154,7 +123,7 @@ mod test {
         let expected_endpoint = Endpoint {
             protocol: "https".to_string(),
             host: "some.test.domain".to_string(),
-            port: MIDEN_NODE_PORT,
+            port: Endpoint::MIDEN_NODE_PORT,
         };
 
         assert_eq!(endpoint, expected_endpoint);
@@ -166,7 +135,7 @@ mod test {
         let expected_endpoint = Endpoint {
             protocol: "https".to_string(),
             host: "192.168.0.1".to_string(),
-            port: MIDEN_NODE_PORT,
+            port: Endpoint::MIDEN_NODE_PORT,
         };
 
         assert_eq!(endpoint, expected_endpoint);
@@ -202,7 +171,7 @@ mod test {
         let expected_endpoint = Endpoint {
             protocol: "hkttp".to_string(),
             host: "some.test.domain".to_string(),
-            port: MIDEN_NODE_PORT,
+            port: Endpoint::MIDEN_NODE_PORT,
         };
 
         assert_eq!(endpoint, expected_endpoint);
@@ -214,7 +183,7 @@ mod test {
         let expected_endpoint = Endpoint {
             protocol: "http".to_string(),
             host: "192.168.0.1".to_string(),
-            port: MIDEN_NODE_PORT,
+            port: Endpoint::MIDEN_NODE_PORT,
         };
 
         assert_eq!(endpoint, expected_endpoint);