+++ crates/rust-client/src/store/sqlite_store/mod.rs
@@ -3,11 +3,11 @@ use alloc::{
     collections::{BTreeMap, BTreeSet},
     vec::Vec,
 };
-use std::{path::Path, string::ToString};
+use std::{path::PathBuf, string::ToString};
 
 use deadpool_sqlite::{Config, Hook, HookError, Pool, Runtime};
 use miden_objects::{
-    accounts::{Account, AccountHeader, AccountId, AuthSecretKey},
+    accounts::{Account, AccountCode, AccountHeader, AccountId, AuthSecretKey},
     crypto::merkle::{InOrderIndex, MmrPeaks},
     notes::{NoteTag, Nullifier},
     BlockHeader, Digest, Word,
@@ -15,7 +15,6 @@ use miden_objects::{
 use rusqlite::{vtab::array, Connection};
 use tonic::async_trait;
 
-use self::config::SqliteStoreConfig;
 use super::{
     ChainMmrNodeFilter, InputNoteRecord, NoteFilter, OutputNoteRecord, Store, TransactionFilter,
 };
@@ -27,7 +26,6 @@ use crate::{
 
 mod accounts;
 mod chain_data;
-pub mod config;
 mod errors;
 mod notes;
 mod sync;
@@ -49,10 +47,10 @@ impl SqliteStore {
     // --------------------------------------------------------------------------------------------
 
     /// Returns a new instance of [Store] instantiated with the specified configuration options.
-    pub async fn new(config: &SqliteStoreConfig) -> Result<Self, StoreError> {
-        let database_exists = Path::new(&config.database_filepath).exists();
+    pub async fn new(database_filepath: PathBuf) -> Result<Self, StoreError> {
+        let database_exists = database_filepath.exists();
 
-        let connection_cfg = Config::new(config.database_filepath.clone());
+        let connection_cfg = Config::new(database_filepath);
         let pool = connection_cfg
             .builder(Runtime::Tokio1)
             .map_err(|err| StoreError::DatabaseError(err.to_string()))?
@@ -295,6 +293,27 @@ impl Store for SqliteStore {
             .await
     }
 
+    async fn upsert_foreign_account_code(
+        &self,
+        account_id: AccountId,
+        code: AccountCode,
+    ) -> Result<(), StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::upsert_foreign_account_code(conn, account_id, code)
+        })
+        .await
+    }
+
+    async fn get_foreign_account_code(
+        &self,
+        account_ids: Vec<AccountId>,
+    ) -> Result<BTreeMap<AccountId, AccountCode>, StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::get_foreign_account_code(conn, account_ids)
+        })
+        .await
+    }
+
     async fn get_unspent_input_note_nullifiers(&self) -> Result<Vec<Nullifier>, StoreError> {
         self.interact_with_connection(SqliteStore::get_unspent_input_note_nullifiers)
             .await
@@ -306,18 +325,10 @@ impl Store for SqliteStore {
 
 #[cfg(test)]
 pub mod tests {
-    use std::string::ToString;
-
-    use super::{config::SqliteStoreConfig, SqliteStore};
+    use super::SqliteStore;
     use crate::mock::create_test_store_path;
 
     pub(crate) async fn create_test_store() -> SqliteStore {
-        let temp_file = create_test_store_path();
-
-        SqliteStore::new(&SqliteStoreConfig {
-            database_filepath: temp_file.to_string_lossy().to_string(),
-        })
-        .await
-        .unwrap()
+        SqliteStore::new(create_test_store_path()).await.unwrap()
     }
 }