+++ crates/rust-client/src/store/web_store/accounts/js_bindings.rs
@@ -72,4 +72,14 @@ extern "C" {
         auth_info: Vec<u8>,
         pub_key: Vec<u8>,
     ) -> js_sys::Promise;
+
+    #[wasm_bindgen(js_name = upsertForeignAccountCode)]
+    pub fn idxdb_upsert_foreign_account_code(
+        account_id: String,
+        code: Vec<u8>,
+        code_root: String,
+    ) -> js_sys::Promise;
+
+    #[wasm_bindgen(js_name = getForeignAccountCode)]
+    pub fn idxdb_get_foreign_account_code(account_ids: Vec<String>) -> js_sys::Promise;
 }